+++ cxx_mkspecs/clang_mkspecs.py
@@ -206,6 +206,19 @@ def cxx_apple_llvm90_x64(conf):
 
 
 @conf
+def cxx_apple_llvm91_x64(conf):
+    """
+    Detect and setup the 64-bit Apple LLVM 9.1 compiler
+    """
+    if conf.is_mkspec_platform('mac'):
+        conf.mkspec_clang_configure(9, 1)
+        conf.mkspec_add_common_flag('-m64')
+    else:
+        conf.fatal("This mkspec is not supported on {0}.".format(
+            conf.get_mkspec_platform()))
+
+
+@conf
 def cxx_clang36_x64(conf):
     """
     Detect and setup the clang 3.6 compiler for 64 bit